Rename move and split float-clamp.h to passes/FloatTrap.(h|cpp)
